refactor: move ci-github command inside subfolder
